refractor: move styles to theme folder
