+++ lib/widgets/navigation/bottom_nav_bar.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:staugustinechsnewapp/styles.dart';
+import 'package:staugustinechsnewapp/theme/styles.dart';
 import 'package:staugustinechsnewapp/utilities/navigation/enav_index_conversion.dart';
 import 'package:staugustinechsnewapp/utilities/navigation/nav_bloc.dart';
