+++ lib/widgets/reusable/basic_container.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:staugustinechsnewapp/styles.dart';
+import 'package:staugustinechsnewapp/theme/styles.dart';
 
 /// {@template basic_container}
 /// Reusable widget for displaying a container with a border radius and shadow.
